Refactored all calls to get chart data into ChartService
