+++ api/src/main/java/org/pmiops/workbench/cohortbuilder/CohortBuilderServiceImpl.java
@@ -40,10 +40,8 @@ import org.pmiops.workbench.cdr.model.DbCriteria;
 import org.pmiops.workbench.cdr.model.DbCriteriaAttribute;
 import org.pmiops.workbench.cohortbuilder.mapper.CohortBuilderMapper;
 import org.pmiops.workbench.db.model.DbConceptSetConceptId;
-import org.pmiops.workbench.model.AgeType;
 import org.pmiops.workbench.model.AgeTypeCount;
 import org.pmiops.workbench.model.CardCount;
-import org.pmiops.workbench.model.CohortChartData;
 import org.pmiops.workbench.model.ConceptIdName;
 import org.pmiops.workbench.model.Criteria;
 import org.pmiops.workbench.model.CriteriaAttribute;
@@ -51,12 +49,9 @@ import org.pmiops.workbench.model.CriteriaListWithCountResponse;
 import org.pmiops.workbench.model.CriteriaMenu;
 import org.pmiops.workbench.model.CriteriaType;
 import org.pmiops.workbench.model.DataFilter;
-import org.pmiops.workbench.model.DemoChartInfo;
 import org.pmiops.workbench.model.Domain;
 import org.pmiops.workbench.model.DomainCard;
-import org.pmiops.workbench.model.EthnicityInfo;
 import org.pmiops.workbench.model.FilterColumns;
-import org.pmiops.workbench.model.GenderOrSexType;
 import org.pmiops.workbench.model.ParticipantDemographics;
 import org.pmiops.workbench.model.SearchRequest;
 import org.pmiops.workbench.model.SurveyModule;
@@ -217,27 +212,6 @@ public class CohortBuilderServiceImpl implements CohortBuilderService {
     return personDao.findAgeTypeCounts().stream()
         .map(cohortBuilderMapper::dbModelToClient)
         .collect(Collectors.toList());
-  }
-
-  @Override
-  public List<CohortChartData> findCohortChartData(
-      SearchRequest searchRequest, Domain domain, int limit) {
-    TableResult result =
-        bigQueryService.executeQuery(
-            bigQueryService.filterBigQueryConfig(
-                cohortQueryBuilder.buildDomainChartInfoCounterQuery(
-                    new ParticipantCriteria(searchRequest), domain, limit)));
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-
-    List<CohortChartData> cohortChartData = new ArrayList<>();
-    for (List<FieldValue> row : result.iterateAll()) {
-      cohortChartData.add(
-          new CohortChartData()
-              .name(bigQueryService.getString(row, rm.get("name")))
-              .conceptId(bigQueryService.getLong(row, rm.get("conceptId")))
-              .count(bigQueryService.getLong(row, rm.get("count"))));
-    }
-    return cohortChartData;
   }
 
   @Override
@@ -423,46 +397,6 @@ public class CohortBuilderServiceImpl implements CohortBuilderService {
     return StreamSupport.stream(cbDataFilterDao.findAll().spliterator(), false)
         .map(cohortBuilderMapper::dbModelToClient)
         .collect(Collectors.toList());
-  }
-
-  @Override
-  public List<DemoChartInfo> findDemoChartInfo(
-      GenderOrSexType genderOrSexType, AgeType ageType, SearchRequest request) {
-    QueryJobConfiguration qjc =
-        bigQueryService.filterBigQueryConfig(
-            cohortQueryBuilder.buildDemoChartInfoCounterQuery(
-                new ParticipantCriteria(request, genderOrSexType, ageType)));
-    TableResult result = bigQueryService.executeQuery(qjc);
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-
-    List<DemoChartInfo> demoChartInfos = new ArrayList<>();
-    for (List<FieldValue> row : result.iterateAll()) {
-      demoChartInfos.add(
-          new DemoChartInfo()
-              .name(bigQueryService.getString(row, rm.get("name")))
-              .race(bigQueryService.getString(row, rm.get("race")))
-              .ageRange(bigQueryService.getString(row, rm.get("ageRange")))
-              .count(bigQueryService.getLong(row, rm.get("count"))));
-    }
-    return demoChartInfos;
-  }
-
-  @Override
-  public List<EthnicityInfo> findEthnicityInfo(SearchRequest request) {
-    QueryJobConfiguration qjc =
-        bigQueryService.filterBigQueryConfig(
-            cohortQueryBuilder.buildEthnicityInfoCounterQuery(new ParticipantCriteria(request)));
-    TableResult result = bigQueryService.executeQuery(qjc);
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-
-    List<EthnicityInfo> ethnicityInfos = new ArrayList<>();
-    for (List<FieldValue> row : result.iterateAll()) {
-      ethnicityInfos.add(
-          new EthnicityInfo()
-              .ethnicity(bigQueryService.getString(row, rm.get("ethnicity")))
-              .count(bigQueryService.getLong(row, rm.get("count"))));
-    }
-    return ethnicityInfos;
   }
 
   @Override
